+++ zikzak_inappwebview_linux/linux/include/zikzak_inappwebview_linux/in_app_webview.h
@@ -15,10 +15,6 @@ InAppWebView* in_app_webview_new(FlBinaryMessenger* messenger, FlTextureRegistra
 // Registers initial user scripts and loads the initial URL/data from
 // the "params" argument map of the create/run channel calls.
 void in_app_webview_load_initial(InAppWebView* self, FlValue* params);
-
-// Registers initial user scripts and loads the initial URL/data from
-// the "params" argument map of the create/run channel calls.
-void in_app_webview_load_initial(InAppWebView* self, FlValue* params);
 int64_t in_app_webview_get_texture_id(InAppWebView* self);
 void in_app_webview_handle_method_call(InAppWebView* self, FlMethodCall* method_call);